// ui/Enums/AbbreviationsEnum.ts
export enum PermissionsShortValueEnum {
    ÜSZİ = "Ücretsiz İzin",
    Üİ = "Ücretli İzin",
    Yİ = "Yıllık İzin",
    R = "Raporlu",
    HST = "Hafta Sonu Tatili",
    RT = "Resmi Tatil",
    V = "Vardiya"
}
export enum PermissionsShortKeyEnum {
    Ücretsiz_İzin = "ÜSZİ",
    Ücretli_İzin = "Üİ",
    Yıllık_İzin = "Yİ",
    Raporlu = "R",
    Hafta_Sonu_Tatili = "HST",
    Resmi_Tatil = "RT",
    Vardiya ="V"
}
export enum DaysOfWeekShortKeyEnumTR {
    PZT = "Pazartesi",
    S = "Salı",
    ÇRŞ = "Çarşamba",
    PRŞ = "Perşembe",
    C = "Cuma",
    CRT = "Cumartesi",
    P = "Pazar"
}
export enum DaysOfWeekShortValueEnumEN {
    PZT = "monday",
    S = "tuesday",
    ÇRŞ = "wednesday",
    PRŞ = "thursday",
    C = "friday",
    CRT = "saturday",
    P = "sunday"
}
export enum DaysOfWeekShortKeyEnumEN {
    monday = "PZT",
    tuesday = "S",
    wednesday = "ÇRŞ",
    thursday = "PRŞ",
    friday = "C",
    saturday = "CRT",
    sunday = "P"
}